Extract `NearestStopsMapper` type to its own file
